requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0.0
openai>=1.0.0
pymupdf>=1.24.0
pypdf>=4.0.0
